+++ packages/cli/CHANGELOG.md
@@ -17,7 +17,7 @@
 
 ### Patch Changes
 
-- 24878ed: Use @guild-docs/mdx-remote with esbuild as peer dependency (no need for esbuild resolutions/overrides)
+- 24878ed: Use @pablosz/guild-docs-mdx-remote with esbuild as peer dependency (no need for esbuild resolutions/overrides)
 
 ## 1.0.1