+++ packages/mdx-remote/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @guild-docs/mdx-remote
+# @pablosz/guild-docs-mdx-remote
 
 ## 1.0.2
 
@@ -16,9 +16,9 @@
 
 ### Major Changes
 
-- 24878ed: Use @guild-docs/mdx-remote with esbuild as peer dependency (no need for esbuild resolutions/overrides)
+- 24878ed: Use @pablosz/guild-docs-mdx-remote with esbuild as peer dependency (no need for esbuild resolutions/overrides)
 
 ### Patch Changes
 
 - Updated dependencies [24878ed]
-  - @guild-docs/types@1.0.3
+  - @pablosz/guild-docs-types@1.0.3